Updates from review
