+++ readthedocs/support/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import url
 
-from readthedocs.support.views import FrontWebhook
+from readthedocs.support.views import FrontAppWebhook
 
 urlpatterns = [
-    url(r'^front-webhook/$', FrontWebhook.as_view(), name='front_webhook'),
+    url(r'^frontapp-webhook/$', FrontAppWebhook.as_view(), name='frontapp_webhook'),
 ]
